gt - SortableTable Header to TS
Also update required props
